Update banner tests
