+++ clients/privacy-center/cypress/support/mocks.ts
@@ -3,6 +3,7 @@ import {
   EnforcementLevel,
   ConsentMechanism,
   UserConsentPreference,
+  FidesCookie,
 } from "fides-js";
 
 export const mockPrivacyNotice = (params: Partial<PrivacyNotice>) => {
@@ -33,4 +34,22 @@ export const mockPrivacyNotice = (params: Partial<PrivacyNotice>) => {
     cookies: [],
   };
   return { ...notice, ...params };
+};
+
+export const mockCookie = (params: Partial<FidesCookie>) => {
+  const uuid = "4fbb6edf-34f6-4717-a6f1-541fd1e5d585";
+  const CREATED_DATE = "2022-12-24T12:00:00.000Z";
+  const UPDATED_DATE = "2022-12-25T12:00:00.000Z";
+  const cookie: FidesCookie = {
+    identity: { fides_user_device_id: uuid },
+    fides_meta: {
+      version: "0.9.0",
+      createdAt: CREATED_DATE,
+      updatedAt: UPDATED_DATE,
+    },
+    consent: {},
+    tcf_consent: {},
+  };
+
+  return { ...cookie, ...params };
 };
